requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
delegator.py==0.1.1
distlib==0.3.1
Django==3.1.3
Django-Basic-CRUD-Generator==0.2
django-celery-beat==2.1.0
django-dbbackup==3.3.0
django-timezone-field==4.0
docopt==0.6.2
filelock==3.0.12
idna==2.10
Js2Py==0.70
kombu==5.0.2
packaging==20.4
pexpect==4.8.0
pipwin==0.5.0
prompt-toolkit==3.0.8
psycopg2 @ file:///C:/Users/%D0%90%D0%BD%D0%B4%D1%80%D0%B5%D0%B9/pipwin/psycopg2-2.8.6-cp39-cp39-win_amd64.whl
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.4
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.25.11
vine==5.0.0
virtualenv==20.1.0
virtualenv-clone==0.5.4
wcwidth==0.2.5
